src/requirement.txt
pandas
numpy
scipy.optimize
matplotlib.pyplot
seaborn
typing
datetime
dateutil.relativedelta